Separate recorder options and interceptor
